% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr2-package.R
\docType{package}
\name{samplr2-package}
\alias{samplr2}
\alias{samplr2-package}
\title{samplr2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/samplr2}
  \item Report bugs at \url{https://github.com/paezha/samplr2/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
